Pulls the money types into their own package for sharing Renames transactionType to transactionDirection for clarity Adds a broken starter implementaiton of the moveCashBetweenPots function
