+++ functions/moveCashBetweenPots/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"encoding/json"
+	"sync"
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
@@ -13,6 +14,14 @@ import (
 	"theodo.red/creditcompanion/packages/credtrack"
 	"theodo.red/creditcompanion/packages/logging"
 )
+
+func ProcessTransaction(logger logging.Logger, clientRepo clirepo.ClientRepository, transaction credtrack.CreditTransaction) {
+	var waitGroup sync.WaitGroup
+
+	for clientId := transaction.LinkedClients {
+
+	}
+}
 
 func handleRequest(ctx context.Context, e events.DynamoDBEvent) {
 	sess := session.Must(session.NewSession())
