+++ packages/credtrack/go.mod
@@ -1,3 +1,7 @@
 module theodo.red/creditcompanion/packages/credtrack
 
 go 1.15
+
+replace theodo.red/creditcompanion/packages/money => ../money
+
+require theodo.red/creditcompanion/packages/money v1.0.0
