--- packages/credtrack/package.go
@@ -1,1 +1,0 @@
-package credtrack
