+++ packages/money/go.mod
@@ -1,0 +1,5 @@
+module theodo.red/creditcompanion/packages/money
+
+go 1.15
+
+require github.com/stretchr/testify v1.7.0
