+++ packages/pot/go.mod
@@ -1,0 +1,3 @@
+module theodo.red/creditcompanion/packages/pot
+
+go 1.15
